requirements.txt
boto3==1.26.114
botocore==1.29.114
ecdsa==0.18.0
jmespath==1.0.1
pyasn1==0.4.8
python-dateutil==2.8.2
python-jose==3.3.0
rsa==4.9
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.15
